requirements.txt
asgiref==3.8.1
crispy-bootstrap3==2024.1
Django==5.0.7
django-appconf==1.0.6
django-crispy-forms==2.2
django-image-cropping==1.7
django-imagekit==5.0.0
django-widget-tweaks==1.5.0
easy-thumbnails==2.8.5
pilkit==3.0
pillow==10.4.0
sqlparse==0.5.1
tzdata==2024.1
whitenoise==6.7.0
